refactor: 타입스크립트 마이그레이션, 타입가드(is) 및 로직 안정화(옵셔널)
